Update Test Results doc (#219)
In a pester v5 upgrade I noticed that you needed to enable `TestResults` in the new configuration object to get pester to actually create a file. I could not find this mentioned anywhere except in an obscure example https://pester.dev/docs/commands/Invoke-Pester#example-4. Thought it belonged in the Test Result docs